[fix] #95 빌드 에러 해결(delayList, focusList LAZY 접근 문제)

+++ src/main/java/dgu/umc_app/domain/plan/dto/request/PlanDelayRequest.java
@@ -23,6 +23,6 @@ public record PlanDelayRequest(
 
         @Schema(description = "수행 시간")
         @NotNull
-        Long executeTime
+        int executeTime
 ) {}
